feat(wasm): allow for optimization passes

+++ crates/core/src/compile_wasm.rs
@@ -1,15 +1,16 @@
 use indicatif::ProgressBar;
 use std::{
     borrow::Cow,
-    fs::{self, File},
-    io::{self, Write},
+    ffi::OsStr,
+    fs, io,
     path::{Path, PathBuf},
     process::Command,
     str,
     time::Duration,
 };
 
-use anyhow::{bail, Context, Error, Result};
+use anyhow::{anyhow, bail, Context, Error, Result};
+use binaryen::{CodegenConfig, Module};
 use decorous_backend::{dom_render::DomRenderer, prerender::Prerenderer, CodeInfo, WasmCompiler};
 use itertools::Itertools;
 use scopeguard::defer;
@@ -17,6 +18,7 @@ use shlex::Shlex;
 use which::which;
 
 use crate::{
+    cli::OptimizationLevel,
     config::{Config, ScriptOrFile},
     FINISHED,
 };
@@ -26,14 +28,21 @@ pub struct MainCompiler<'a> {
     config: &'a Config,
     build_args: &'a [(String, String)],
     out_name: &'a str,
+    opt_level: Option<OptimizationLevel>,
 }
 
 impl<'a> MainCompiler<'a> {
-    pub fn new(config: &'a Config, out_name: &'a str, build_args: &'a [(String, String)]) -> Self {
+    pub fn new(
+        config: &'a Config,
+        out_name: &'a str,
+        build_args: &'a [(String, String)],
+        opt_level: Option<OptimizationLevel>,
+    ) -> Self {
         Self {
             config,
             build_args,
             out_name,
+            opt_level,
         }
     }
 
@@ -79,16 +88,14 @@ macro_rules! compile_for {
                 let spinner = ProgressBar::new_spinner().with_message(format!("Building WebAssembly ({lang})..."));
                 spinner.enable_steady_tick(Duration::from_micros(100));
 
-                {
-                    let mut f = File::create(&path)?;
-                    f.write_all(body.as_bytes())?;
-                }
+                fs::write(&path, body)?;
 
-                let _guard = scopeguard::guard(&path, |p| {
-                    fs::remove_file(p).unwrap_or_else(|_| {
-                        panic!("error removing \"{}\"! Remove it manually!", p.display())
+                defer! {
+                    // No .expect() to save on format! call
+                    fs::remove_file(&path).unwrap_or_else(|_| {
+                        panic!("error removing \"{}\"! Remove it manually!", path.display())
                     });
-                });
+                }
 
                 match fs::create_dir(self.out_name) {
                     Ok(()) => {}
@@ -96,38 +103,32 @@ macro_rules! compile_for {
                     Err(err) => bail!(err),
                 }
 
-                let python = self.get_python().context("python not found in PATH! Make sure to install it!")?;
+                let python = self.get_python().context("python not found in $PATH! Make sure to install it!")?;
                 let mut build_args = self.build_args_for_lang(lang);
 
-                let (status, stdout, stderr) = match &config.script {
-                    ScriptOrFile::File(file) => {
-                        let out = Command::new(python.as_ref())
-                            .arg(file)
-                            .env("DECOR_INPUT", &path)
-                            .env("DECOR_OUT", self.out_name)
-                            .env("DECOR_EXPORTS", exports.iter().join(" "))
-                            .args(&mut build_args)
-                            .output()?;
-                        (out.status, out.stdout, out.stderr)
-                    }
+                let file_loc = match &config.script {
+                    ScriptOrFile::File(file) => file.as_path(),
                     ScriptOrFile::Script(script) => {
-                        {
-                            let mut f = File::create("__tmp.py")?;
-                            f.write_all(script.as_bytes())?;
-                        }
-                        defer! {
-                            fs::remove_file("__tmp.py").expect("error removing \"__tmp.py\"! Remove it manually!");
-                        }
-                        let out = Command::new(python.as_ref())
-                            .arg("__tmp.py")
-                            .env("DECOR_INPUT", &path)
-                            .env("DECOR_OUT", self.out_name)
-                            .env("DECOR_EXPORTS", exports.iter().join(" "))
-                            .args(&mut build_args)
-                            .output()?;
-                        (out.status, out.stdout, out.stderr)
-                    }
+                        fs::write("__tmp.py", script)?;
+                        Path::new("__tmp.py")
+                    },
                 };
+                // This defer! cannot be used in the above match statement, as it executes when a
+                // scope ends, and match arms have individual scopes
+                defer! {
+                    if matches!(&config.script, ScriptOrFile::Script(_)) {
+                        fs::remove_file("__tmp.py").expect("error removing \"__tmp.py\"! Remove it manually!");
+                    }
+                }
+
+                let script_out = Command::new(python.as_ref())
+                    .arg(file_loc)
+                    .env("DECOR_INPUT", &path)
+                    .env("DECOR_OUT", self.out_name)
+                    .env("DECOR_EXPORTS", exports.iter().join(" "))
+                    .args(&mut build_args)
+                    .output()?;
+                let (status, stdout, stderr) = (script_out.status, script_out.stdout, script_out.stderr);
 
                 if build_args.had_error {
                     bail!("error parsing build args for language: {lang}");
@@ -139,6 +140,35 @@ macro_rules! compile_for {
                         str::from_utf8(&stderr)?,
                         str::from_utf8(&stdout)?,
                     );
+                }
+
+
+                if let Some(opt) = self.opt_level {
+                    let (shrink, speed) = match opt {
+                        OptimizationLevel::SpeedMinor => (0, 1),
+                        OptimizationLevel::SpeedMedium => (0, 2),
+                        OptimizationLevel::SpeedMajor => (0, 3),
+                        OptimizationLevel::SpeedAggressive => (0, 4),
+                        OptimizationLevel::Size => (1, 2),
+                        OptimizationLevel::SizeAggressive => (2, 2),
+                    };
+                    for entry in fs::read_dir(self.out_name)? {
+                        let entry = entry?;
+                        let path = entry.path();
+
+                        match path.extension() {
+                            Some(ext) if ext == OsStr::new("wasm") => {},
+                            _ => continue,
+                        }
+
+                        let contents = fs::read(&path)?;
+                        // Uses wasm-opt (https://github.com/WebAssembly/binaryen) optimizations
+                        let mut module = Module::read(&contents).map_err(|_err| anyhow!("could not optimize .wasm file: {}", path.display()))?;
+                        let config = CodegenConfig { shrink_level : shrink, optimization_level: speed, debug_info: false };
+                        module.optimize(&config);
+                        let out = module.write();
+                        fs::write(&path, out)?;
+                    }
                 }
 
                 out.write_all(&stdout)?;
